requirements.txt
pillow>=10.2.0
numpy>=1.26.4
